+++ src/main/java/com/rnt/test_passing/exception/TestReadingException.java
@@ -1,0 +1,11 @@
+package com.rnt.test_passing.exception;
+
+public class TestReadingException extends RuntimeException {
+  public TestReadingException(String message, Throwable cause){
+    super(message, cause);
+  }
+
+  public TestReadingException(String message){
+    super(message);
+  }
+}
