+++ src/main/java/com/rnt/test_passing/repo/QuestionRepository.java
@@ -1,0 +1,12 @@
+package com.rnt.test_passing.repo;
+
+import java.util.List;
+import com.rnt.test_passing.entity.Question;
+import com.rnt.test_passing.exception.TestReadingException;
+
+public interface QuestionRepository {
+  List<Question> findQuestionsByName(String name) throws TestReadingException;
+
+  List<String> getQuestionTopicNames() throws TestReadingException;
+
+}
