+++ src/main/java/com/rnt/test_passing/service/IOService.java
@@ -1,0 +1,8 @@
+package com.rnt.test_passing.service;
+
+public interface IOService {
+
+  void printText(String text);
+
+  String readText();
+}
