+++ src/main/java/com/rnt/test_passing/service/QuestionService.java
@@ -1,0 +1,10 @@
+package com.rnt.test_passing.service;
+
+import java.util.List;
+import com.rnt.test_passing.entity.Question;
+
+public interface QuestionService {
+
+  List<Question> getQuestions(String testName);
+
+}
