+++ src/main/java/com/rnt/test_passing/service/TestExecutor.java
@@ -1,0 +1,9 @@
+package com.rnt.test_passing.service;
+
+import java.util.List;
+import com.rnt.test_passing.entity.Question;
+
+public interface TestExecutor {
+  void startTest(List<Question> testList);
+
+}
