+++ src/main/java/com/rnt/test_passing/service/TestService.java
@@ -1,0 +1,7 @@
+package com.rnt.test_passing.service;
+
+import java.util.List;
+
+public interface TestService {
+  List<String> getTestNames();
+}
