+++ src/main/java/com/rnt/test_passing/util/QuestionConverter.java
@@ -1,0 +1,7 @@
+package com.rnt.test_passing.util;
+
+import com.rnt.test_passing.entity.Question;
+
+public interface QuestionConverter {
+  String getAnswerAsString(Question question);
+}
