--- src/main/java/org/main/entity/OptionCatalogue.java
@@ -1,14 +1,0 @@
-package org.main.entity;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class OptionCatalogue {
-
-  private Map<String, Boolean> options = new HashMap<>();
-
-  public Map<String, Boolean> getOptions() {
-    return options;
-  }
-
-}
