--- src/main/java/org/main/exception/TestReadingException.java
@@ -1,11 +1,0 @@
-package org.main.exception;
-
-public class TestReadingException extends Throwable {
-  public TestReadingException(String message, Throwable cause){
-    super(message, cause);
-  }
-
-  public TestReadingException(String message){
-    super(message);
-  }
-}
