--- src/main/java/org/main/service/IOService.java
@@ -1,8 +1,0 @@
-package org.main.service;
-
-public interface IOService {
-
-  void printText(String text);
-
-  String nextString();
-}
