--- src/main/java/org/main/service/QuestionService.java
@@ -1,13 +1,0 @@
-package org.main.service;
-
-import java.util.List;
-import org.main.entity.Question;
-import org.main.entity.Result;
-
-public interface QuestionService {
-
-  List<Question> getQuestions(String testName);
-
-  List<String> getTestNames();
-
-}
