--- src/main/java/org/main/service/TestService.java
@@ -1,9 +1,0 @@
-package org.main.service;
-
-import java.util.List;
-import org.main.entity.Question;
-
-public interface TestService {
-  void startTest(List<Question> testList);
-
-}
