+++ src/main/resources/application-context.xml
@@ -3,30 +3,40 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-  <bean id="questionRepository" class="org.main.repo.impl.QuestionRepositoryImpl">
-    <constructor-arg ref="pathProvider"/>
+  <bean id="sourceFileDescriptorHelperImpl" class="com.rnt.test_passing.util.impl.SourceFileDescriptorHelperImpl">
+    <constructor-arg name="basePaths" value="${application.path}"/>
   </bean>
 
-  <bean id="testService" class="org.main.service.impl.TestServiceImpl">
+  <bean id="questionRepository" class="com.rnt.test_passing.repo.impl.QuestionRepositoryImpl">
+    <constructor-arg ref="sourceFileDescriptorHelperImpl"/>
+  </bean>
+
+  <bean id="testService" class="com.rnt.test_passing.service.impl.TestServiceImpl">
+    <constructor-arg ref="questionRepository"/>
     <constructor-arg ref="ioService"/>
   </bean>
 
-  <bean id="questionService" class="org.main.service.impl.QuestionServiceImpl">
+  <bean id="testExecutor" class="com.rnt.test_passing.service.impl.TestExecutorImpl">
+    <constructor-arg ref="ioService"/>
+    <constructor-arg ref="questionConverter"/>
+  </bean>
+
+  <bean id="questionService" class="com.rnt.test_passing.service.impl.QuestionServiceImpl">
     <constructor-arg ref="ioService"/>
     <constructor-arg ref="questionRepository"/>
   </bean>
 
-  <bean id="pathProvider" class="org.main.util.PathProvider">
-    <constructor-arg name="basePaths" value="${application.path}"/>
+  <bean id="questionConverter" class="com.rnt.test_passing.util.impl.QuestionConverterImpl">
   </bean>
 
-  <bean id="ioService" class="org.main.service.impl.IOServiceImpl">
+  <bean id="ioService" class="com.rnt.test_passing.service.impl.IOServiceImpl">
   </bean>
 
-  <bean id="appService" class="org.main.service.AppService">
-    <constructor-arg ref="testService"/>
+  <bean id="appService" class="com.rnt.test_passing.service.AppService">
+    <constructor-arg ref="testExecutor"/>
     <constructor-arg ref="questionService"/>
     <constructor-arg ref="ioService"/>
+    <constructor-arg ref="testService"/>
   </bean>
 
   <bean id="conversionService"
